// src/app/models/InvoiceHeaderDetail.model.ts
export class InvoiceHeaderDetail {
      HEADER_ID: number;
      PLANT: string;
      PLANT_NAME: string;
      INV_NO: string;
      INV_DATE: string;
      INV_TYPE: string;
      CUSTOMER: string;
      CUSTOMER_NAME: string;
      ODIN: number;
      VEHICLE_NO: string;
      VEHICLE_CAPACITY: string;
      POD_DATE: string;
      EWAYBILL_NO: string;
      EWAYBILL_DATE: string;
      LR_NO: string;
      LR_DATE: string;
      FWD_AGENT: string;
      CARRIER: string;
      FREIGHT_ORDER: number;
      FREIGHT_ORDER_DATE: string;
      OUTBOUND_DELIVERY: string;
      OUTBOUND_DELIVERY_DATE: string;
      ACTUAL_DISPATCH_DATE: string;
      PROPOSED_DELIVERY_DATE: string;
      VEHICLE_REPORTED_DATE: string;
      ACTUAL_DELIVERY_DATE: string;
      TRANSIT_LEAD_TIME: string;
      CANC_INV_STATUS: string;
      STATUS: string;
      STATUS_DESCRIPTION: string;
      ISXMLCREATED: boolean;
      XMLMOVED_ON: string;
      CREATED_BY: string;
      CREATED_ON: string;
      IS_ACTIVE: boolean;
      ATTACHMENT_ID:number;
      ATTACHMENT_NAME:string;

}